done CRUD cate and lab

+++ src/main/java/com/vncodelab/respository/LabRespository.java
@@ -10,13 +10,13 @@ import org.springframework.stereotype.Repository;
 import com.vncodelab.entity.Lab;
 
 @Repository
-public interface LabRespository extends JpaRepository<Lab, Long> {
+public interface LabRespository extends JpaRepository<Lab, Integer> {
 
 	Lab findByDocID(String docID);
 
 	boolean existsByDocID(String docID);
 
 	List<Lab> findAllByCateID(int cateID);
-	
+
 	Page<Lab> findAll(Pageable pageable);
 }
